backend/requirements.txt
fastapi==0.100.0
openai==0.27.8
pydantic==1.10.9
python-dotenv==1.0.0
uvicorn==0.22.0
